+++ logging/src/main/java/net/daporkchop/lib/logging/impl/DefaultLogger.java
@@ -1,0 +1,10 @@
+package net.daporkchop.lib.logging.impl;
+
+/**
+ * @author DaPorkchop_
+ */
+public class DefaultLogger extends BaseLogger {
+    public DefaultLogger()  {
+        super(System.out::println);
+    }
+}
